Add LauncherService.kt to the shortcut package
